% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_TWRI_logo.R
\name{add_TWRI_logo}
\alias{add_TWRI_logo}
\title{Add logo to ggplot}
\usage{
add_TWRI_logo(plot, scale = 0.25, halign = 0.01, valign = 0.01)
}
\arguments{
\item{plot}{ggplot object}

\item{scale}{scale the logo, defaults to 0.25}

\item{halign}{aligns logo within the ggplot [0-1]}

\item{valign}{aligns the logo within the ggplot [0-1]}
}
\value{
ggplot object
}
\description{
Add logo to ggplot
}
