% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dissolved_oxygen}
\alias{dissolved_oxygen}
\title{Dissolved oxygen measurements from the Tres Palacios river}
\format{
A data frame with 236 rows and 6 variables:
\describe{
\item{station_id}{unique water quality monitoring station identifier}
\item{end_date}{sampling date in yyyy-mm-dd format}
\item{parameter_code}{unique parameter code}
\item{parameter_description}{parameter description with units}
\item{average_do}{mean of dissolved oxygen measurement, in mg/L}
\item{min_do}{minimum of dissolved oxygen measurement, in mg/L}
}
}
\source{
\url{https://www80.tceq.texas.gov/SwqmisPublic/public/default.htm}
}
\usage{
data(dissolved_oxygen)
}
\description{
Data from the Texas Commission on Environmental Quality Surface Water Quality
Monitoring Information System. The \verb{Average_DO`` field is the mean of dissolved oxygen concentrations (mg/L) measured at a field site at that day. The }Min_DO` field
is the minimum dissolved oxygen concentration measured at that site on that day.
}
\keyword{datasets}
