% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{easterwood_weather}
\alias{easterwood_weather}
\title{Temperature and precipitation data at Easterwood Airport}
\format{
A data frame with 4045 rows and 5 variables:
\describe{
\item{station}{station identifier}
\item{date}{sampling date in yyyy-mm-dd format}
\item{name}{station name}
\item{dailymaximumdrybulbtemperature}{daily maximum tempreature in degrees F.}
\item{dailyprecipitation}{daily total precipitation in inches}
}
}
\source{
https://www.ncei.noaa.gov/cdo-web/datasets/GHCND/stations/GHCND:USW00003904/detail
}
\usage{
data(easterwood_weather)
}
\description{
Data from the National Oceanic Atmospheric Administration (NOAA) Climate
Data Online for Easterwood Airport in College Station, Texas. Includes
daily maximum temperature and daily precipitation records from approximately
2010 through 2017.
}
\keyword{datasets}
