% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_palette.R
\name{get_twri_palettes}
\alias{get_twri_palettes}
\title{Get names of all unique palettes provided in twriTemplates}
\usage{
get_twri_palettes(full = FALSE)
}
\arguments{
\item{full}{Whether to include full palette names (with suffixes, e.g. \code{_cont}) or just stubs}
}
\value{
Vector of palette name stubs or full names
}
\description{
Get names of all unique palettes provided in twriTemplates
}
\examples{
get_twri_palettes()
}
