% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_box_legend.R
\name{ggplot_box_legend}
\alias{ggplot_box_legend}
\title{Create a Boxplot Legend}
\usage{
ggplot_box_legend(
  font_family = "OpenSansCondensed_TWRI",
  point_label = " - Measured values"
)
}
\arguments{
\item{font_family}{defaults to \code{"OpenSansCondensed_TWRI"}. Can be
modified to any locally installed font.}

\item{point_label}{character string - how you want to label to point in a
legend. If you overlay measured points in the box plot, \code{" - Measured
  values"} is appropriate. If the points only indicate outliers, \code{" -
  values outside the\\ninterquartile range"} is appropriate. This can be
custoimized as necessary. Use \code{"\\n"} to indicate where a line break
should occur.}
}
\value{
a ggplot object
}
\description{
Create a legend that explains the features of a boxplot. Can be added easily
to existing ggplots using \code{\link[cowplot]{plot_grid}} or \code{\link[patchwork]{plot_arithmetic}}.
}
\examples{
library(ggplot2)
library(cowplot)
ggplot(Dissolved_Oxygen) +
  geom_boxplot(aes(as.factor(Station_ID),
                   Average_DO,
                   group = Station_ID)) +
  geom_jitter(aes(as.factor(Station_ID),
                  Average_DO,
                  group = Station_ID),
               width = 0.1, alpha = 0.25) +
 theme_TWRI_print() -> p1
file <- tempfile(fileext = '.png')
ragg::agg_png(file)
plot_grid(p1, ggplot_box_legend(), rel_widths = c(2,1))
dev.off()
}
\references{
\url{https://waterdata.usgs.gov/blog/boxplots/}
}
\author{
Laura DeCicco
}
