% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_font.R
\name{install_font}
\alias{install_font}
\title{Open and install fonts}
\usage{
install_font(font)
}
\arguments{
\item{font}{one of \code{c("CrimsonPro","OpenSansCondensed")}}
}
\value{
Hopefully nothing
}
\description{
Helper function to open font folders in your GUI so you can install fonts
for local use.
}
\examples{
install_font("CrimsonPro")
}
