% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mission_aransas_nerr}
\alias{mission_aransas_nerr}
\title{Water quality data for Mission and Aransas National Estuarine Research Reserve (NERR)}
\format{
A data frame with 140352 rows and 10 variables:
\describe{
\item{StationCode}{station identifier, factor with levels \code{marabwq} and \code{marcewq} for Aransas Bay and Copano East stations respectively.}
\item{DateTimeStamp}{sampling date-time in yyyy-mm-dd  hh:mm::ss format.}
\item{Temp}{water temperature in degrees Celsius.}
\item{F_Temp}{data qualifier, <0> indicates approved data.}
\item{SpCond}{specific conductance in µS/cm.}
\item{F_SpCond}{data qualifier, <0> indicates approved data.}
\item{Sal}{Salinity in ppt.}
\item{F_Sal}{data qualifier, <0> indicates approved data.}
\item{DO_mgl}{dissolved oxygen in mg/L.}
\item{F_DO_mgl}{data qualifier, <0> indicates approved data.}
}
}
\source{
NOAA National Estuarine Research Reserve System (NERRS). 2022. System-Wide Monitoring Program. NOAA NERRS Centralized Data Management Office. \url{http://cdmo.baruch.sc.edu/}.
}
\usage{
data(mission_aransas_nerr)
}
\description{
Fifteen minute water quality data. Includes station, date-time, temperature,
specific conductance, salinity, and associated
data qualifiers from 2020 through 2021 for the Mission and Aransas (MAR)
National Estuarine Research Reserve (NERR). Data from NOAA NERRS Centralized
Data Management Office.
}
\keyword{datasets}
