% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neon_stage_discharge}
\alias{neon_stage_discharge}
\title{Discharge measurements from field-based surveys}
\format{
A data frame with 134 rows and 36 variables. Primary variables of interest include:
\describe{
\item{averageVelocityUnits}{}
\item{averageVelocityUnitsQF}{}
\item{collectDate}{}
\item{collectedBy}{}
\item{dataQF}{}
\item{dischargeUnitsQF}{}
\item{domainID}{}
\item{filterParamTime}{}
\item{finalDischarge}{Corrected measured stream discharge}
\item{flowCalcQF}{}
\item{flowCalculation}{}
\item{handheldDeviceID}{}
\item{lowVelocityFinalQF}{}
\item{namedLocation}{}
\item{profileName}{}
\item{publicationDate}{}
\item{recorduid}{}
\item{release}{}
\item{samplingProtocolVersion}{}
\item{siteID}{}
\item{stageImpractical}{}
\item{startDate}{}
\item{stationEntryTest}{}
\item{streamStage}{Measured stream stage in meters}
\item{streamStageUnits}{}
\item{streamStageUnitsQF}{}
\item{tapeDistanceUnits}{}
\item{tapeDistanceUnitsQF}{}
\item{totalDischarge}{}
\item{totalDischargeCalcQF}{}
\item{totalDischargeUnits}{}
\item{uid}{}
\item{velocitySensorID}{}
\item{waterDepthUnits}{}
\item{waterDepthUnitsQF}{}
\item{waterEdge}{}
}
}
\source{
NEON (National Ecological Observatory Network). n.d. Discharge Field Collection (DP1.20048.001), RELEASE-2022. \doi{10.48443/eaak-tt31}.
}
\usage{
data(neon_stage_discharge)
}
\description{
Includes field based stream discharge and stage measurements at the Como Creek
NEON field station.
}
\keyword{datasets}
