% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to example data}
\source{
This function is adapted from \code{readxl::readxl_example()}
}
\usage{
path_to_file(file = NULL)
}
\arguments{
\item{file}{Name of the file in quotes with extension; \code{"arroyo_swqm.txt"},
\code{"dissolved_oxygen_raw.csv"}, \code{"easterwood.csv"}, \code{"marabwq2021.csv"}, or
\code{"marcewq2021.csv"} will work. If \code{NULL}, the example files will be listed.}
}
\value{
file path
}
\description{
twriTemplates comes with five bundled  example data files in its
\code{inst/extdata} directory. This function makes them easy to access for
practice reading with \code{\link[=read.csv]{read.csv()}} or \code{\link[readr:read_delim]{readr::read_csv()}}.
}
\examples{
path_to_file()
path_to_file("easterwood.csv")
head(read.csv(path_to_file("easterwood.csv")))
}
