% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_palette.R
\name{scale_custom}
\alias{scale_custom}
\alias{scale_color_discrete_twri}
\alias{scale_colour_discrete_twri}
\alias{scale_fill_discrete_twri}
\alias{scale_color_opinionated_twri}
\alias{scale_colour_opinionated_twri}
\alias{scale_fill_opinionated_twri}
\alias{scale_color_diverging_twri}
\alias{scale_colour_diverging_twri}
\alias{scale_fill_diverging_twri}
\alias{scale_color_continuous_twri}
\alias{scale_colour_continuous_twri}
\alias{scale_fill_continuous_twri}
\title{Custom color and fill scales}
\usage{
scale_color_discrete_twri(palette = "twri", extend = FALSE, ...)

scale_colour_discrete_twri(palette = "twri", extend = FALSE, ...)

scale_fill_discrete_twri(palette = "twri", extend = FALSE, ...)

scale_color_opinionated_twri(palette = "twri", ...)

scale_colour_opinionated_twri(palette = "twri", extend = FALSE, ...)

scale_fill_opinionated_twri(palette = "twri", ...)

scale_color_diverging_twri(palette = "twri", ...)

scale_colour_diverging_twri(palette = "twri", ...)

scale_fill_diverging_twri(palette = "twri", ...)

scale_color_continuous_twri(palette = "twri", ...)

scale_colour_continuous_twri(palette = "twri", ...)

scale_fill_continuous_twri(palette = "twri", ...)
}
\arguments{
\item{palette}{Name of color palette}

\item{extend}{Whether to extend discrete color palette to make sufficient colors for levels needed}

\item{...}{Additional arguments to be passed to internal scale function}
}
\description{
Custom coloring and filling functions based on unique color palettes
}
\details{
Specific functions include:
\itemize{
\item{scale_(color/colour/fill)\emph{discrete_twri}{Discrete palette with either fixed or dynamically extended number of shades}
\item{scale}(color/colour/fill)\emph{opinionated_twri}{Discrete palette with specific values for "good", "bad", and "neutral"}
\item{scale}(color/colour/fill)\emph{diverging_twri}{Continuous diverging color palette}
\item{scale}(color/colour/fill)_continuous}{Continuous color palette}
}
}
