% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_TWRI_fig.R
\name{theme_TWRI_fig}
\alias{theme_TWRI_fig}
\title{a [ggplot2] theme formatted in the TWRI style}
\usage{
theme_TWRI_fig(
  base_size = 8.5,
  base_family = "Arial",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{default text size in px}

\item{base_family}{defaults to \code{"Arial"}. Generally, use a sensible sans serif font that is available on your system.}

\item{base_line_size}{default line size}

\item{base_rect_size}{default rect size}
}
\description{
Provides a [ggplot2] theme formatted consistent with TWRI style guidance.
}
