% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_TWRI_pres.R
\name{theme_TWRI_pres}
\alias{theme_TWRI_pres}
\title{a [ggplot2] theme formatted in the TWRI style for presentations}
\usage{
theme_TWRI_pres(
  base_size = 16,
  base_family = "Oswald_TWRI",
  base_line_size = 1,
  base_rect_size = 1
)
}
\arguments{
\item{base_size}{default text size in px}

\item{base_family}{defaults to \code{"Moriston_personal"}. Generally, use a sensible sans serif font that is available on your system.}

\item{base_line_size}{default line size}

\item{base_rect_size}{default rect size}
}
\description{
Provides a [ggplot2] theme formatted consistent with TWRI style guidance. Figure aesthetics prioritize design and legibility for use in power point presentations. The intended output is 10 inches by 7.5 inches as 200dpi.
}
