% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_rmd.R
\name{twri_docx}
\alias{twri_docx}
\title{RMD Word Template for TWRI}
\usage{
twri_docx(
  base_format = "bookdown::word_document2",
  toc = FALSE,
  number_sections = TRUE,
  fig_captions = TRUE,
  reference_docx = "format.docx",
  ...
)
}
\arguments{
\item{base_format}{character, either "rmarkdown::word_document" or
"bookdown::word_document2" (default)}

\item{toc}{logical defaults FALSE. The template utilizes the TOC function in
Word that can be manually removed or customized.}

\item{number_sections}{logical}

\item{fig_captions}{logical}

\item{reference_docx}{character, specifies the template document to use.}

\item{...}{additional arguments to \@code{rdocx_document}}
}
\description{
Loads an Rmarkdown template that will produce a Microsoft word document
consistent with institute brand guidance.
}
