% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_palette.R
\docType{data}
\name{twri_pal}
\alias{twri_pal}
\alias{twri_pal_op}
\alias{twri_pal_cont}
\alias{twri_pal_div}
\title{TWRI complementary color palette}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
twri_pal

twri_pal_op

twri_pal_cont

twri_pal_div
}
\description{
This is a test palette
}
\keyword{datasets}
