% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_rmd.R
\name{twri_pdf}
\alias{twri_pdf}
\title{RMD PDF Template for TWRI}
\usage{
twri_pdf(
  toc = TRUE,
  toc_depth = 2,
  number_sections = TRUE,
  latex_engine = "lualatex",
  ...
)
}
\arguments{
\item{toc}{Logical}

\item{toc_depth}{numeric}

\item{number_sections}{logical}

\item{latex_engine}{character, one of "pdflatex", "lualatex", or "xelatex"}

\item{...}{arguments passed to pdf_document}
}
\description{
Loads an Rmarkdown template that will produce a pdf document
consistent with institute brand guidance.
}
