% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twriTemplates.R
\docType{package}
\name{twriTemplates-package}
\alias{twriTemplates}
\alias{twriTemplates-package}
\title{twriTemplates: Templates for TWRI reports}
\description{
The twriTemplates package provides rmarkdown templates that follow internal brand guidance. It also provides functions for formatting ggplot2 plots that follow brand guides.
}
\author{
\strong{Maintainer}: Michael Schramm \email{michael.schramm@ag.tamu.edu} (\href{https://orcid.org/0000-0003-1876-6592}{ORCID})

}
