% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uranium_tds}
\alias{uranium_tds}
\title{Uranium and total dissolved solids}
\format{
A data frame with 44 rows and 4 variables:
\describe{
\item{tds}{Total dissolved solids in mg/L}
\item{uranium}{Uranium concentration in ppb}
\item{hco3}{cation composition percent milliequivalents}
\item{definition}{not documented}
}
}
\source{
Helsel, D.R., Hirsch, R.M., Ryberg, K.R., Archfield, S.A., and Gilroy, E.J., 2020, Statistical Methods in Water Resources - Supporting Materials: U.S. Geological Survey data release, \doi{10.5066/P9JWL6XR}
}
\usage{
data(uranium_tds)
}
\description{
Example dataset from Helsel et al. (2020) that includes
measured concentrations of total dissolved solids (tds) in mg/L and uranium
in parts per billion from the Ogallala aquifer.
}
\keyword{datasets}
