% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neon_stage_discharge}
\alias{neon_stage_discharge}
\title{Discharge measurements from field-based surveys}
\format{
A data frame with 134 rows and 36 variables. Primary variables of interest include:
\describe{
\item{streamStage}{Measured stream stage in meters}
\item{finalDischarge}{Corrected measured stream discharge}
}
}
\source{
NEON (National Ecological Observatory Network). n.d. Discharge Field Collection (DP1.20048.001), RELEASE-2022. doi.org/10.48443/eaak-tt31.
}
\usage{
data(neon_stage_discharge)
}
\description{
Includes field based stream discharge and stage measurements at the Como Creek
NEON field station.
}
\keyword{datasets}
