% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twri_rmd.R
\name{twri_docx}
\alias{twri_docx}
\title{RMD docx template for TWRI}
\usage{
twri_docx(
  base_format = "bookdown::word_document2",
  tables = list(),
  plots = list(),
  lists = list(),
  mapstyles = list(),
  page_size = list(),
  page_margins = list(),
  reference_num = TRUE,
  toc = FALSE,
  number_sections = TRUE,
  fig_captions = TRUE,
  reference_docx = "format.docx",
  ...
)
}
\arguments{
\item{base_format}{character, either \code{rmarkdown::word_document} or
\code{bookdown::word_document2} (default)}

\item{tables}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{plots}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{lists}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{mapstyles}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{page_size}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{page_margins}{empty list, used by \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}

\item{reference_num}{if \code{TRUE} text for references to sections will be the
section number. If \code{FALSE} the text for references to sections will be the
section title. Default is \code{TRUE}.}

\item{toc}{logical defaults FALSE. The template utilizes the TOC function in
Word that can be manually removed or customized.}

\item{number_sections}{logical}

\item{fig_captions}{logical}

\item{reference_docx}{character, specifies the template document to use.}

\item{...}{additional arguments to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Development on \code{twri_docx()} is complete. It is recommended to use
\code{officedown::rdocx_document()} directly. This function is retained for older
templates that include the function, although rendered documents might be
inconsistent. It is suggested to update any reports to the latest version of
twriTemplates as \code{twri_docx()} may be removed in future versions.

Loads an Rmarkdown template that will produce a Microsoft word document
consistent with institute brand guidance.
}
